Ex. 4.21.

+++ part4/blog_list/controllers/blogs.js
@@ -34,9 +34,24 @@ blogsRouter.post('/', async (request, response, next) => {
   }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  const result = await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', async (request, response, next) => {
+  try {
+    const result = await Blog.findById(request.params.id)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    const user = await User.findById(decodedToken.id)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    if (result.user.toString() === user.id.toString()) {
+      const result = await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).json(result)
+    }
+    else {
+      return response.status(401).json({ error: 'only blog creator can delete' })
+    }
+  } catch(exception) {
+    next(exception)
+  }
 })
 
 blogsRouter.patch('/:id', async (request, response) => {
